### 2018-04-18 #680 @whedon
[PRE REVIEW]: ParticleScattering: Solving and optimizing multiple-scattering problems in Julia
**Submitting author:** @bblankrot (<a href="http://orcid.org/0000-0003-3364-9298">Boaz Blankrot</a>)
**Repository:** <a href="https://github.com/bblankrot/ParticleScattering.jl" target ="_blank">https://github.com/bblankrot/ParticleScattering.jl</a>
**Version:** v0.0.2
