### 2018-04-18 #680 @bblankrot
Thanks, proof looks good overall.

From the list of reviewers, I suggest any of the following:
- ChrisRackauckas
- ysimillides
- ziotom78
- lheagy
