### 2018-04-20 d8846f2
Moving from plane wave to arbitrary fields
1st commit
