+++ examples/sim2_LuneburgOptim.jl
@@ -43,21 +43,21 @@ border = (R_lens + a_lens)*[-1;1;-1;1]
 
 sp1 = ScatteringProblem([CircleParams(rs_lnbrg[i]) for i in eachindex(rs_lnbrg)],
         ids_lnbrg, centers, φs)
-Ez1 = plot_near_field(k0, kin, P, sp1, θ_i, x_points = 150, y_points = 150,
+Ez1 = plot_near_field(k0, kin, P, sp1, PlaneWave(θ_i), x_points = 150, y_points = 150,
         opt = fmm_options, border = border)
 plot_near_field_pgf(filename1, k0, kin, P, sp1, θ_i; opt = fmm_options,
     x_points = 201, y_points = 201, border = border)
 
 sp2 = ScatteringProblem([CircleParams(rs_max[i]) for i in eachindex(rs_max)],
         ids_max, centers, φs)
-Ez2 = plot_near_field(k0, kin, P, sp2, θ_i, x_points = 150, y_points = 150,
+Ez2 = plot_near_field(k0, kin, P, sp2, PlaneWave(θ_i), x_points = 150, y_points = 150,
             opt = fmm_options, border = border)
 plot_near_field_pgf(filename2, k0, kin, P, sp2, θ_i; opt = fmm_options,
     x_points = 201, y_points = 201, border = border)
 
 sp3 = ScatteringProblem([CircleParams(rs0[i]) for i in eachindex(rs0)],
         collect(1:length(rs0)), centers, φs)
-Ez3 = plot_near_field(k0, kin, P, sp3, θ_i, x_points = 150, y_points = 150,
+Ez3 = plot_near_field(k0, kin, P, sp3, PlaneWave(θ_i), x_points = 150, y_points = 150,
         opt = fmm_options, border = border)
 plot_near_field_pgf(filename3, k0, kin, P, sp3, θ_i; opt = fmm_options,
     x_points = 201, y_points = 201, border = border)
@@ -122,18 +122,18 @@ JLD.@save output_dir * "/luneburg_optim_sym.jld" test_max_sym sym_time
 
 sp4 = ScatteringProblem([CircleParams(rs_sym[i]) for i in eachindex(rs_sym)],
         ids_sym, centers, φs)
-Ez4 = plot_near_field(k0, kin, P, sp4, θ_i, x_points = 150, y_points = 150,
+Ez4 = plot_near_field(k0, kin, P, sp4, PlaneWave(θ_i), x_points = 150, y_points = 150,
         opt = fmm_options, border = border)
-
-u1 = calc_near_field(k0, kin, 7, sp1, points, θ_i; opt = fmm_options)
-u2 = calc_near_field(k0, kin, 7, sp2, points, θ_i; opt = fmm_options)
-u3 = calc_near_field(k0, kin, 7, sp3, points, θ_i; opt = fmm_options)
-u4 = calc_near_field(k0, kin, 7, sp4, points, θ_i; opt = fmm_options)
+pw = PlaneWave(θ_i)
+u1 = calc_near_field(k0, kin, 7, sp1, points, pw; opt = fmm_options)
+u2 = calc_near_field(k0, kin, 7, sp2, points, pw; opt = fmm_options)
+u3 = calc_near_field(k0, kin, 7, sp3, points, pw; opt = fmm_options)
+u4 = calc_near_field(k0, kin, 7, sp4, points, pw; opt = fmm_options)
 abs.([u1[1];u2[1];u3[1];u4[1]])
 
 #####################################
 # selfconsistent err P calculation
[…]
