+++ src/multipole.jl
@@ -1,9 +1,9 @@
 """
-	solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, θ_i = 0.0; get_inner = true, verbose = true) -> beta, inner
+	solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, pw::PlaneWave; get_inner = true, verbose = true) -> beta, inner
 
 Solve the scattering problem `sp` with outer wavenumber `k0`, inner wavenumber
 `kin`, `2P+1` cylindrical harmonics per inclusion and incident plane wave angle
-`θ_i`. Solves multiple-scattering equation directly.
+`pw.θi`. Solves multiple-scattering equation directly.
 Returns the cylindrical harmonics basis `beta` along with potential
 densities (in case of arbitrary inclusion) or inner cylindrical coefficients (in
 case of circular). By default, incident wave propagates left->right.
@@ -11,7 +11,7 @@ case of circular). By default, incident wave propagates left->right.
 Inner coefficients are only calculated if `get_inner` is true, and timing is
 printed if `verbose` is true.
 """
-function solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, θ_i = 0.0;
+function solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, pw::PlaneWave;
 								get_inner = true, verbose = true)
 	# This function solves for the outgoing multipole coefficients in the presence
 	# of an incident plane wave.
@@ -20,13 +20,12 @@ function solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, θ_i = 0.0
 	shapes = sp.shapes;	ids = sp.ids; centers = sp.centers; φs = sp.φs
 	Ns = size(sp)
 	#first solve for single scatterer densities
-	α = Complex{Float64}[exp(1.0im*p*(pi/2-θ_i)) for p=-P:P]
 	tic()
 	scatteringMatrices,innerExpansions = particleExpansion(k0, kin, shapes, P, ids)
 	dt1 = toq()
 	tic()
 	T = Array{Complex{Float64}}(Ns*(2*P+1),Ns*(2*P+1))
-	a = Array{Complex{Float64}}(Ns*(2*P+1))
+	a = u2α(k0, pw, centers, P)
 	Btemp = Array{Complex{Float64}}(2*P+1,2*P+1)
 	for ic1 = 1:Ns
 		rng1 = (ic1-1)*(2*P+1) + (1:2*P+1)
@@ -45,8 +44,77 @@ function solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, θ_i = 0.0
 				T[rng1,rng2] = -RotScatMat*Btemp
 			end
 		end
-		phase = exp(1.0im*k0*(cos(θ_i)*centers[ic1,1] + sin(θ_i)*centers[ic1,2])) #phase shift added to move cylinder coords
-		a[rng1] = phase*(RotScatMat*α)
+		a[rng1] = RotScatMat*a[rng1]
+	end
+	dt2 = toq()
+	tic()
+	beta = T\a
+	dt3 = toq()
+	#recover full incoming expansion - in sigma_mu terms for parametrized shape,
+	#in multipole expansion for circle
+	if get_inner
+		tic()
+		#find LU factorization once for each shape
+        scatteringLU = [lufact(scatteringMatrices[iid]) for iid = 1:length(shapes)]
[…]
