### 2018-04-20 ca42aaf
Moving from plane wave to arbitrary fields
2nd commit
