+++ examples/sim3_AngleOpt.jl
@@ -43,7 +43,7 @@ begin
     if draw_fig
         figure()
         #draw shapes and points
-        draw_shapes(shapes, centers, ids, φs)
+        draw_shapes(shapes, ids, centers, φs)
         plot(points[:,1], points[:,2], "r*")
         tight_layout()
         ax = gca()
@@ -75,7 +75,7 @@ plot_near_field(k0, kin, P, sp_before, PlaneWave(θ_i),
 colorbar()
 clim([0;5])
 plot_near_field_pgf(output_dir * "/opt_phi_before.tex", k0, kin, P,
-    sp_before, θ_i; opt = fmm_options, x_points = 150, y_points = 50,
+    sp_before, PlaneWave(θ_i); opt = fmm_options, x_points = 150, y_points = 50,
     border = plot_border, downsample = 10, include_preamble = true)
 
 sp_after = ScatteringProblem(shapes, ids, centers, test_max.minimizer)
@@ -84,7 +84,7 @@ plot_near_field(k0, kin, P, sp_after, PlaneWave(θ_i),
 colorbar()
 clim([0;5])
 plot_near_field_pgf(output_dir * "/opt_phi_after.tex", k0, kin, P,
-    sp_after, θ_i; opt = fmm_options, x_points = 150, y_points = 50,
+    sp_after, PlaneWave(θ_i); opt = fmm_options, x_points = 150, y_points = 50,
     border = plot_border, downsample = 10, include_preamble = true)
 
 inner_iters = length(test_max.trace)
