+++ src/visualization_pgf.jl
@@ -1,11 +1,11 @@
 """
-    plot_near_field_pgf(filename, k0, kin, P, sp::ScatteringProblem, θ_i = 0;
+    plot_near_field_pgf(filename, k0, kin, P, sp::ScatteringProblem, pw::PlaneWave;
                         opt::FMMoptions = FMMoptions(), use_multipole = true,
                         x_points = 201, y_points = 201, border = find_border(sp),
                         downsample = 1, include_preamble = false, normalize = 1.0)
 
 Plots the total electric field as a result of a plane wave with incident
-angle `θ_i` scattering from the ScatteringProblem `sp`, using pgfplots's
+angle `pw.θi` scattering from the ScatteringProblem `sp`, using pgfplots's
 `surf`. Can accept number of sampling points in each direction, and either
 a given `border` or calculate it automatically. The plots of the shapes (but not
 the field) can be downsampled by setting an integer `downsample`, since pgfplots
@@ -21,7 +21,7 @@ Saves the generated pgfplots file to `filename`, with just a surrounding `tikzpi
 environment if `include_preamble=false`, and a compilable tandalone document
 otherwise.
 """
-function plot_near_field_pgf(filename, k0, kin, P, sp::ScatteringProblem, θ_i = 0;
+function plot_near_field_pgf(filename, k0, kin, P, sp::ScatteringProblem, pw::PlaneWave;
                             opt::FMMoptions = FMMoptions(), use_multipole = true,
                             x_points = 201, y_points = 201, border = find_border(sp),
                             downsample = 1, include_preamble = false, normalize = 1.0)
@@ -36,7 +36,7 @@ function plot_near_field_pgf(filename, k0, kin, P, sp::ScatteringProblem, θ_i =
     ygrid = repmat(y,1,x_points)
     points = [xgrid[:] ygrid[:]]
 
-    Ez = calc_near_field(k0, kin, P, sp, points, PlaneWave(θ_i),
+    Ez = calc_near_field(k0, kin, P, sp, points, pw,
                             use_multipole=use_multipole, opt = opt)
 
     aspect = (y_max - y_min)/(x_max - x_min)
@@ -62,11 +62,11 @@ function plot_near_field_pgf(filename, k0, kin, P, sp::ScatteringProblem, θ_i =
                     mesh/rows={$(y_points)}] table{$(basename(filename)).dat};")
 
     end
-    draw_shapes_pgf(shapes, centers, ids, φs, ax, 1.1*maximum(abs.(Ez)), downsample, normalize = normalize)
+    draw_shapes_pgf(shapes, ids, centers, φs, ax, 1.1*maximum(abs.(Ez)), downsample, normalize = normalize)
     pgf.save(filename, ax ,include_preamble = include_preamble)
 end
 
-function draw_shapes_pgf(shapes, centers, ids, φs, ax, floating, downsample; normalize = 1.0)
+function draw_shapes_pgf(shapes, ids, centers, φs, ax, floating, downsample; normalize = 1.0)
     #draw in 3d so it is "above" surf plot
     for ic = 1:size(centers,1)
         if typeof(shapes[ids[ic]]) == ShapeParams
