### 2018-04-23 a5c11d8
Remove unneeded type specification
