### 2018-04-23 #691 @ziotom78
Everything looks fine to me, I created two PRs ([#18](https://github.com/bblankrot/ParticleScattering.jl/pull/18) and [#19](https://github.com/bblankrot/ParticleScattering.jl/pull/19)) to improve the documentation and make the plots actually appear.

I have checked the claim in the paper that the optimization example takes 35 seconds to complete: on my (quite old) laptop the two calls to `optimize_φ` take 27 s and 18 s, so this point is ok as well.
