### 2018-04-23 c3c9057
typo in calc_far_field
