### 2018-04-23 #691 @bblankrot
Thanks for reviewing this so quickly, @ziotom78 , as well as fixing the issues! I merged both PRs.
