+++ latest/search_index.js
@@ -53,7 +53,7 @@ var documenterSearchIndex = {"docs": [
     "page": "Tutorial 1: Solving a Simple Multiple-Scattering Problem",
     "title": "Scattering from one particle",
     "category": "section",
-    "text": "In order to speed up computations, ParticleScattering depends on grouping identical shapes (rotated versions of the same shape are considered identical). Thus solving multiple-scattering problems is syntactically similar to solving for a single particle.In this example, we simulate TM plane-wave scattering (E_z = e^ik(cos theta_i  sin theta_i) cdot mathbfr) from a single rounded star, parametrized by the equation(x(theta) y(theta)) = R + d cos(5theta)(cos theta sin theta) quad\nR = 01lambda_0 d = 005lambda_0which is supplied by rounded_star. For now, we discretize the shape with N=260 nodes and P=10 cylindrical harmonics &ndash; for more information on the relationship between these parameters and the various resulting errors, see Choosing Minimal N and P.λ0 = 1 #doesn\'t matter since everything is normalized to λ0\nk0 = 2π/λ0\nkin = 3k0\nθ_i = 0.0 #incident wave is left->right\n\nN = 260\nP = 10\nshapes = [rounded_star(0.1λ0, 0.05λ0, 5, N)]\nids = [1] # the particle at centers[1,:] has the parametrization shapes[ids[1]]\ncenters = [0.0 0.0] # our particle is centered at the origin\nφs = [0.0] #zero rotation angle\nsp = ScatteringProblem(shapes, ids, centers, φs)Now that the scattering problem is set up, we solve for the cylindrical harmonics coefficients and potential densities, respectively, by usingbeta,inner = solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, θ_i)These can be used to calculate the scattered field at any point in space using low-level function scatteredfield, or strictly outside the circle of radius shapes[1].R with scattered_field_multipole. For large numbers of calculation points, however, it is easier to use calc_near_field which performs solve_particle_scattering and calculates the total (incident + scattered) field at every point using the most appropriate method:#calculate field on the x-axis passing through the particle\npoints = [linspace(-0.5λ0, 0.5λ0, 200)  zeros(200)]\nu = calc_near_field(k0, kin, P, sp, points, θ_i)\nplot(points[:,1]/λ0, abs.(u))(Image: simple_tutorial_plot1)Similarly, a 2D plot can be drawn of the total field around the scatterer:plot_near_field(k0, kin, P, sp::ScatteringProblem, θ_i;\n    x_points = 201, y_points = 201, border = 0.5λ0*[-1;1;-1;1])(Image: simple_tutorial_plot2)Note: In practice, converting the shape potential densities to cylindrical harmonics is inefficient here as we only have one scatterer, and get_potentialPW would be more accurate. This is meant only as an introductory example to the ParticleScattering syntax."
[…]
