### 2018-04-23 7984cf8
Moving from plane wave to arbitrary fields
3rd commit
