+++ src/multipole.jl
@@ -1,9 +1,9 @@
 """
-	solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, pw::PlaneWave; get_inner = true, verbose = true) -> beta, inner
+	solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, u::Einc; get_inner = true, verbose = true) -> beta, inner
 
 Solve the scattering problem `sp` with outer wavenumber `k0`, inner wavenumber
-`kin`, `2P+1` cylindrical harmonics per inclusion and incident plane wave angle
-`pw.θi`. Solves multiple-scattering equation directly.
+`kin`, `2P+1` cylindrical harmonics per inclusion and incident TM field
+`u`. Solves multiple-scattering equation directly.
 Returns the cylindrical harmonics basis `beta` along with potential
 densities (in case of arbitrary inclusion) or inner cylindrical coefficients (in
 case of circular). By default, incident wave propagates left->right.
@@ -11,12 +11,8 @@ case of circular). By default, incident wave propagates left->right.
 Inner coefficients are only calculated if `get_inner` is true, and timing is
 printed if `verbose` is true.
 """
-function solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, pw::PlaneWave;
+function solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, u::Einc;
 								get_inner = true, verbose = true)
-	# This function solves for the outgoing multipole coefficients in the presence
-	# of an incident plane wave.
-	# incident wave direction - from left to right is 0:
-	# e^{ik(\cos(\θ_i),\sin(\θ_i)) \cdot \mathbf{r}}
 	shapes = sp.shapes;	ids = sp.ids; centers = sp.centers; φs = sp.φs
 	Ns = size(sp)
 	#first solve for single scatterer densities
@@ -25,77 +21,7 @@ function solve_particle_scattering(k0, kin, P, sp::ScatteringProblem, pw::PlaneW
 	dt1 = toq()
 	tic()
 	T = Array{Complex{Float64}}(Ns*(2*P+1),Ns*(2*P+1))
-	a = u2α(k0, pw, centers, P)
-	Btemp = Array{Complex{Float64}}(2*P+1,2*P+1)
-	for ic1 = 1:Ns
-		rng1 = (ic1-1)*(2*P+1) + (1:2*P+1)
-		if φs[ic1] == 0.0 || typeof(shapes[ids[ic1]]) == CircleParams
-			RotScatMat = scatteringMatrices[ids[ic1]]
-		else
-			Rot = spdiagm(Complex{Float64}[exp(-1.0im*φs[ic1]*l) for l=-P:P]) #rotation matrix
-			RotScatMat = Rot*(scatteringMatrices[ids[ic1]]*conj(Rot))
-		end
-		for ic2 = 1:Ns
-			rng2 = (ic2-1)*(2*P+1) + (1:2*P+1)
-			if ic1 == ic2
-				T[rng1,rng2] = eye(Complex{Float64},2*P+1)
-			else
-				M2Lmatrix!(Btemp, k0, P, centers[ic1,:] - centers[ic2,:])
-				T[rng1,rng2] = -RotScatMat*Btemp
-			end
-		end
-		a[rng1] = RotScatMat*a[rng1]
-	end
-	dt2 = toq()
-	tic()
-	beta = T\a
-	dt3 = toq()
-	#recover full incoming expansion - in sigma_mu terms for parametrized shape,
-	#in multipole expansion for circle
-	if get_inner
-		tic()
[…]

[…]
