### 2018-04-24 #691 @Kevin-Mattheus-Moerman
@bblankrot the citation in your paper:
> Blankrot, Boaz, and Clemens Heitzinger. 2018. “Efficient Computational Design and
> Optimization of Dielectric Metamaterial Devices.” _Submitted for Publication_.
Would you be able to upload a pre-print e.g. via [engrxiv.org/](http://engrxiv.org/), and to add the pre-print DOI to the JOSS paper?
