### 2018-04-26 7f31062
Update paper.bib
citing preprint
