+++ paper/paper.bib
@@ -1,8 +1,9 @@
 @article{ar:blankrot2018,
 	author = {Boaz Blankrot and Clemens Heitzinger},
 	title = {Efficient computational design and optimization of dielectric metamaterial devices},
-	journal = {submitted for publication},
-	year = {2018}
+	year = {2018},
+	eprint = {arXiv:1804.09489},
+	url = {https://arxiv.org/abs/1804.09489}
 }
 
 @article{ar:optimjl,
