### 2018-04-26 #691 @bblankrot
@Kevin-Mattheus-Moerman thanks for the comment, I uploaded the preprint to arxiv. Is the citation OK as is or are any changes necessary?
