### 2018-04-27 ef3c9a1
Moving from plane wave to arbitrary fields
4th commit
