+++ src/optimize_rs.jl
@@ -1,5 +1,5 @@
 """
-    optimize_radius(rs0, r_min, r_max, points, ids, P, θ_i, k0, kin, centers,
+    optimize_radius(rs0, r_min, r_max, points, ids, P, ui, k0, kin, centers,
         fmmopts, optimopts::Optim.Options; minimize = true, method = "BFGS")
 
 Optimize the radii of circular particles for minimization or maximization of the
@@ -15,7 +15,7 @@ for both the inner and outer iterations. `method` can be either `"BFGS"` or
 
 Returns an object of type `Optim.MultivariateOptimizationResults`.
 """
-function optimize_radius(rs0, r_min, r_max, points, ids, P, θ_i, k0, kin,
+function optimize_radius(rs0, r_min, r_max, points, ids, P, ui, k0, kin,
                         centers, fmmopts, optimopts::Optim.Options;
                         minimize = true, method = "BFGS")
     Ns = size(centers,1)
@@ -46,7 +46,8 @@ function optimize_radius(rs0, r_min, r_max, points, ids, P, θ_i, k0, kin,
 
     #stuff that is done once
     H = optimizationHmatrix(points, centers, Ns, P, k0)
-    α_inc = Complex{Float64}[exp(1.0im*p*(π/2-θ_i)) for p=-P:P]
+    α = u2α(k0, ui, centers, P)
+    uinc_ = uinc(k0, points, ui)
     φs = zeros(Float64,Ns)
 
     # Allocate buffers
@@ -57,21 +58,21 @@ function optimize_radius(rs0, r_min, r_max, points, ids, P, θ_i, k0, kin,
 
     if minimize
         df = OnceDifferentiable(rs -> optimize_radius_f(rs, last_rs, shared_var,
-                                        φs, α_inc, H, points, P, θ_i, Ns, k0,
+                                        φs, α, H, points, P, uinc_, Ns, k0,
                                         kin, centers,scatteringMatrices, dS_S,
                                         ids, mFMM, fmmopts, buf),
                     (grad_stor, rs) -> optimize_radius_g!(grad_stor, rs, last_rs,
-                                        shared_var, φs, α_inc, H, points, P, θ_i,
+                                        shared_var, φs, α, H, points, P, uinc_,
                                         Ns, k0, kin, centers, scatteringMatrices,
                                         dS_S, ids, mFMM, fmmopts, buf, minimize),
                                         initial_rs)
     else
         df = OnceDifferentiable(rs -> -optimize_radius_f(rs, last_rs, shared_var,
-                                        φs, α_inc, H, points, P, θ_i, Ns, k0,
+                                        φs, α, H, points, P, uinc_, Ns, k0,
                                         kin, centers,scatteringMatrices, dS_S,
                                         ids, mFMM, fmmopts, buf),
                     (grad_stor, rs) -> optimize_radius_g!(grad_stor, rs, last_rs,
-                                        shared_var, φs, α_inc, H, points, P, θ_i,
+                                        shared_var, φs, α, H, points, P, uinc_,
                                         Ns, k0, kin, centers, scatteringMatrices,
                                         dS_S, ids, mFMM, fmmopts, buf, minimize),
                                         initial_rs)
[…]
