### 2018-04-29 #691 @ysimillides
Looks good! Two very minor thing's i'd suggest adding are JLD to the Require file (as you use it in the majority of the examples), and citing the Julia paper as a reference .
