+++ paper/paper.bib
@@ -31,3 +31,16 @@
 	publisher={Nature Research},
 	doi={10.1038/nnano.2015.304}
 }
+
+@article{ar:julia2017,
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal={SIAM Review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  publisher={SIAM},
+  doi={10.1137/141000671},
+  url={https://doi.org/10.1137/141000671}
+}
