### 2018-04-29 #691 @bblankrot
Thank you, @ysimillides! Both good ideas, I'm merging your PR and adding a reference to the 2017 Julia paper in SIAM Review.
