### 2018-05-01 ca6861a
Move to arbitrary fields
- added documentation
- changed testing to use line source
