### 2018-05-05 #691 @Kevin-Mattheus-Moerman
@bblankrot at this point, can you please make an archive of the final reviewed software in [Zenodo](https://zenodo.org/), or [figshare](https://figshare.com/), or another service, and update this thread with the DOI of the archive? Once we have that available we can process acceptance.
