### 2018-05-08 c251519
Faster location tagging, incident field error computation
