### 2018-05-14 #691 @bblankrot
@arfon @Kevin-Mattheus-Moerman - anything else I need to do? The release with the incorporated changes is v0.0.3, archived at https://doi.org/10.5281/zenodo.1241368 .
