### 2018-08-23 ad5a0a6
Updated to Optim 0.15.2
Box constraint optimization has different signature now
