### 2018-08-25 01086e6
updating examples pt 1
