### 2018-09-03 03a397f
calculate only unique ids in optimize_rs
