### 2018-09-05 a7ec09e
Created optimize_φ_adj
