### 2018-09-05 a35055f
updating adjoint optimization over phis, rs
