+++ src/optimize_phis.jl
@@ -6,10 +6,12 @@ Optimize the rotation angles of a particle collection for minimization or
 maximization (depending on `minimize`) of the field intensity at `points`.
 `optimopts` and `method` define the optimization method, convergence criteria,
 and other optimization parameters.
+`adjoint` dictates whether the gradient is calculated using the adjoint method
+(faster) or directly.
 Returns an object of type `Optim.MultivariateOptimizationResults`.
 """
 function optimize_φ(φs0, points, P, ui::Einc, k0, kin, shapes, centers, ids, fmmopts,
-                    optimopts::Optim.Options, method; minimize = true)
+                    optimopts::Optim.Options, method; minimize = true, adjoint = true)
 
     #stuff that is done once
     verify_min_distance(shapes, centers, ids, points) || error("Particles are too close.")
@@ -25,27 +27,22 @@ function optimize_φ(φs0, points, P, ui::Einc, k0, kin, shapes, centers, ids, f
     last_φs = similar(initial_φs)
     last_φs == initial_φs && (last_φs[1] += 1) #should never happen but has
 
-    if minimize
-        df = OnceDifferentiable(
-            φs -> optimize_φ_f(φs, shared_var, last_φs, α,
-                                    H, points, P, uinc_, Ns, k0, centers,
-                                    scatteringMatrices, ids, mFMM, fmmopts, buf),
-            (grad_stor, φs) -> optimize_φ_g!(grad_stor, φs, shared_var,
-                                    last_φs, α, H, points, P, uinc_,
-                                    Ns, k0, centers, scatteringMatrices,
-                                    scatteringLU, ids, mFMM, fmmopts, buf, minimize),
-                                    initial_φs)
+    fopt = φs -> ifelse(minimize,1,-1)*optimize_φ_f(φs, shared_var, last_φs, α,
+                                        H, points, P, uinc_, Ns, k0, centers,
+                                        scatteringMatrices, ids, mFMM, fmmopts,
+                                        buf)
+    if adjoint
+        gopt! = (grad_stor, φs) -> optimize_φ_adj_g!(grad_stor, φs, shared_var,
+                                    last_φs, α, H, points, P, uinc_, Ns, k0,
+                                    centers, scatteringMatrices, scatteringLU,
+                                    ids, mFMM, fmmopts, buf, minimize)
     else
-        df = OnceDifferentiable(
-            φs -> -optimize_φ_f(φs, shared_var, last_φs, α,
-                                    H, points, P, uinc_, Ns, k0, centers,
-                                    scatteringMatrices, ids, mFMM, fmmopts, buf),
-            (grad_stor, φs) -> optimize_φ_g!(grad_stor, φs, shared_var,
-                                    last_φs, α, H, points, P, uinc_,
-                                    Ns, k0, centers, scatteringMatrices,
-                                    scatteringLU, ids, mFMM, fmmopts, buf, minimize),
-                                    initial_φs)
+        gopt! = (grad_stor, φs) -> optimize_φ_g!(grad_stor, φs, shared_var,
+                                    last_φs, α, H, points, P, uinc_, Ns, k0,
+                                    centers, scatteringMatrices, scatteringLU,
+                                    ids, mFMM, fmmopts, buf, minimize)
     end
+    df = OnceDifferentiable(fopt, gopt!, initial_φs)
[…]
