+++ src/optimize_phis_adj.jl
@@ -1,172 +1,0 @@
-"""
-    optimize_φ(φs0, points, P, ui::Einc, k0, kin, shapes, centers, ids, fmmopts,
-        optimopts::Optim.Options, minimize = true)
-
-Optimize the rotation angles of a particle collection for minimization or
-maximization (depending on `minimize`) of the field intensity at `points`.
-`optimopts` and `method` define the optimization method, convergence criteria,
-and other optimization parameters.
-Returns an object of type `Optim.MultivariateOptimizationResults`.
-"""
-function optimize_φ_adj(φs0, points, P, ui::Einc, k0, kin, shapes, centers, ids, fmmopts,
-                    optimopts::Optim.Options, method; minimize = true)
-
-    #stuff that is done once
-    verify_min_distance(shapes, centers, ids, points) || error("Particles are too close.")
-    mFMM, scatteringMatrices, scatteringLU, buf =
-        prepare_fmm_reusal_φs(k0, kin, P, shapes, centers, ids, fmmopts)
-    Ns = size(centers,1)
-    H = optimizationHmatrix(points, centers, Ns, P, k0)
-    α = u2α(k0, ui, centers, P)
-    uinc_ = uinc(k0, points, ui)
-    # Allocate buffers
-    shared_var = OptimBuffer(Ns,P,size(points,1))
-    initial_φs = copy(φs0)
-    last_φs = similar(initial_φs)
-    last_φs == initial_φs && (last_φs[1] += 1) #should never happen but has
-
-    if minimize
-        df = OnceDifferentiable(
-            φs -> optimize_φ_f(φs, shared_var, last_φs, α,
-                                    H, points, P, uinc_, Ns, k0, centers,
-                                    scatteringMatrices, ids, mFMM, fmmopts, buf),
-            (grad_stor, φs) -> optimize_φ_g!(grad_stor, φs, shared_var,
-                                    last_φs, α, H, points, P, uinc_,
-                                    Ns, k0, centers, scatteringMatrices,
-                                    scatteringLU, ids, mFMM, fmmopts, buf, minimize),
-                                    initial_φs)
-    else
-        df = OnceDifferentiable(
-            φs -> -optimize_φ_f(φs, shared_var, last_φs, α,
-                                    H, points, P, uinc_, Ns, k0, centers,
-                                    scatteringMatrices, ids, mFMM, fmmopts, buf),
-            (grad_stor, φs) -> optimize_φ_g!(grad_stor, φs, shared_var,
-                                    last_φs, α, H, points, P, uinc_,
-                                    Ns, k0, centers, scatteringMatrices,
-                                    scatteringLU, ids, mFMM, fmmopts, buf, minimize),
-                                    initial_φs)
-    end
-    optimize(df, initial_φs, method, optimopts)
-end
-
-function optimize_φ_adj_common!(φs, last_φs, shared_var, α, H, points, P, uinc_, Ns, k0, centers, scatteringMatrices, ids, mFMM, opt, buf)
-    if φs != last_φs
-        copy!(last_φs, φs)
-        #do whatever common calculations and save to shared_var
-        #construct rhs
-        for ic = 1:Ns
-            rng = (ic-1)*(2*P+1) + (1:2*P+1)
-            if φs[ic] == 0.0
-                buf.rhs[rng] = scatteringMatrices[ids[ic]]*α[rng]
-            else
[…]
