### 2018-09-08 76ce18d
updated optimize_rs with adj, added gradient calc
