### 2018-10-09 92ea900
Adding multi-frequency optimization
