+++ src/optimize_rs_mf.jl
@@ -1,0 +1,168 @@
+const cₚ = 10.0 #penalty term for unwanted frequency
+
+function optimize_rs_mf(rs0, r_min, r_max, points, point_flags, ids, P, ui, k0,
+                        kin, centers, fmmopts, optimopts::Optim.Options;
+                        method = BFGS(linesearch = LineSearches.BackTracking()))
+
+    Ns = size(centers,1)
+    φs = zeros(Ns)
+    Nk = length(k0); assert(Nk == length(kin))
+    J = length(rs0)
+    assert(maximum(ids) <= J)
+    if length(r_min) == 1
+        r_min = r_min*ones(Float64,J)
+    else
+        assert(J == length(r_min))
+    end
+    if length(r_max) == 1
+        r_max = r_max*ones(Float64,J)
+    else
+        assert(J == length(r_max))
+    end
+    verify_min_distance(CircleParams.(r_max), centers, ids,
+        points) || error("Particles are too close or r_max are too large.")
+
+    groups,boxSize = divideSpace(centers, fmmopts)
+    P2,Q = FMMtruncation(fmmopts.acc, boxSize, maximum(k0))
+    mFMM = [FMMbuildMatrices(k0[ik], P, P2, Q, groups, centers, boxSize) for ik = 1:Nk]
+    scatteringMatrices = [[speye(Complex{Float64}, 2*P+1) for i = 1:J] for ik = 1:Nk]
+    dS_S = [[speye(Complex{Float64}, 2*P+1) for i = 1:J] for ik = 1:Nk]
+
+    #calculate nd expand incident field
+    α = [u2α(k0[ik], ui, centers, P) for ik = 1:Nk]
+    uinc_ = [uinc(k0[ik], points, ui) for ik = 1:Nk]
+    H = [optimizationHmatrix(points, centers, Ns, P, k0[ik]) for ik = 1:Nk]
+
+    # Allocate buffers
+    buf = FMMbuffer(Ns,P,Q,length(groups))
+    shared_var = [OptimBuffer(Ns, P, size(points,1), J) for ik = 1:Nk]
+    initial_rs = copy(rs0)
+    last_rs = similar(initial_rs); last_rs[1] = NaN; assert(last_rs != initial_rs) #initial_rs, last_rs must be different before first iteration
+    df = OnceDifferentiable(
+        rs -> optimize_rs_mf_f(rs, last_rs, shared_var, φs, α, H, points, point_flags,
+                            P, uinc_, Ns, k0, kin, centers, scatteringMatrices, dS_S,
+                            ids, mFMM, fmmopts, buf),
+        (grad_stor, rs) -> optimize_rs_mf_g!(grad_stor, rs, last_rs, shared_var, φs,
+                            α, H, points, point_flags, P, uinc_, Ns, k0, kin, centers,
+                            scatteringMatrices, dS_S, ids, mFMM, fmmopts, buf),
+        initial_rs)
+
+    optimize(df, r_min, r_max, initial_rs, Fminbox(method), optimopts)
+end
+
+function optimize_rs_mf_common!(rs, last_rs, shared_var, φs, α,
+            H, points, P, uinc_, Ns, k0, kin, centers, scatteringMatrices,
+            dS_S, ids, mFMM, fmmopts, buf)
+    if rs != last_rs
+        copy!(last_rs, rs)
+        #do whatever common calculations and save to shared_var
+        #construct rhs
+        for ik in eachindex(k0)
+            for id in unique(ids)
+                try
+                    updateCircleScatteringDerivative!(scatteringMatrices[ik][id],
+                        dS_S[ik][id], k0[ik], kin[ik], rs[id], P)
+                catch
[…]
